requirements.txt
gradio
tensorflow
numpy
Pillow
crewai
langchain
langchain-groq
python-dotenv
langchain_openai
gtts